FreeRTOS-libdl: Do not add per-object globals to the shared/centralised list
Adding each object's globals to the public globals list makes it accessible
for all of the other compartments, a thing that's not desired when we still
want to isolate components. Rather, each object should have its set of lists
(including its globals). When resolving external symbols, we iterate over all
other compartments' interface lists, and iif the policy allows us to access
that particular symol, a symbol is minted/copied to our externals list and
resolved against that.