FreeRTOS-libdl/cheri: Allocate/install a stack per compartment
This allocates a new stack for each compartment and installs it in the very
first slot of the object captable.

We will need to find another solution in the future when we want to support
reentrant compartments by different concurrent tasks. That will need a separate
stack for task per compartment and can be allocated on-demand (e.g., when a task
tries to enter a compartment) or statically. A space vs performance trade-off
arises based on this mechanism used, and we may want to benchmark each and/or
provide them as config options.

Also for this single-stack-per-object, a security concern is if one task entered
a compartment and left, and then another task enters it again; the ISR handler
may need to zero the stack on exit of a compartment to avoid leaking caps and
other auto locals, and this adds a performance overhead.