FreeRTOS-libdl/cheri: Make a capability field an index within a captable
The current capability field in the rtl symbol struct was a pointer to the slot
in the captable for a cap. This wouldn't work when a captable is reallocated
(eg, to make it bigger because it needs to fit more caps).

This changes that behaviour by making the capability field in the symbol struct
a captable index rather than an absolute pointer, and thus there won't be a
problem when the address of the captable changes.