compartmentalization: initialise globals added externally by freertos-syms
Custom prog headers were added to be able to embed and load the symbol table
by ELF loaders that is needed for libdl during dynamic loading and linking. This
works on some targets that blindly load symtables indicated by prog headers,
but not others (such as GDB) that drop loading (in target memory) any sections
having SYMTAB or STRTAB attributes. This is now fixed by
extracting the symbol table and do a 2-pass linking to embed the symbol table
in rodata the same way as RTEMS does, using rtems-tools.

This commit supports using the embedded symbol table by rtems-tools/freertos-syms
instead of assuming the symtable is embedded in the ELF as a separate section.

The rtems-tools/freertos-syms is modified to emit the size and type of each
symbol so that we can create proper bounded caps for it here. Part of this,
the assumption the the sym value is a void* pointer is removed; sym values
are either 32-bit or 64-bit depending on the target, which are different from
pointer types, especially on CHERI.