Do not use the embedded ELF symbol table
An old non-optimal approach used to embed the symbol table in the ELF image
using custom prog headers and output sections. This may not work with all bootloaders
that may skip symtab/strtab by checking section types. Another better solution
is now added which uses an external tool and 2-pass linking to always add the
symbol table and initialise globals, which should work for all bootloaders.

The commit abandon the first old approach and solely relies on the second one.