compartmentalization: Only emit a comp-switch trampoline for inter-compartment calls
Previously all functions within a compartment needed to be wrapped with a trampoline to perform
a compartment switch.

This commit makes use of a new CCALL_GPREL relocation that can be used to detect inter-compartment
calls and passed function pointers. That is, compartmet-switch trampolines are now only emitted
for inter-compartment calls and passed function pointers.